server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.1",
    "annotated-doc>=0.0.4",
    "annotated-types==0.7.0",
    "asyncpg>=0.31.0",
    "audioread>=3.1.0",
    "bcrypt>=5.0.0",
    "black>=25.11.0",
    "certifi>=2025.11.12",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "contourpy>=1.3.3",
    "cryptography>=46.0.3",
    "cycler>=0.12.1",
    "decorator>=5.2.1",
    "demucs>=4.0.1",
    "deprecation>=2.1.0",
    "dnspython>=2.8.0",
    "ecdsa>=0.19.1",
    "email-validator>=2.3.0",
    "exceptiongroup>=1.3.1",
    "fastapi>=0.119.1",
    "fastapi-cli>=0.0.16",
    "fastapi-cloud-cli>=0.5.1",
    "fastar>=0.7.0",
    "fonttools>=4.60.1",
    "greenlet==3.2.4",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "h2>=4.3.0",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httptools>=0.7.1",
    "httpx>=0.28.1",
    "hyperframe>=6.1.0",
    "import-linter>=2.9",
    "isort>=7.0.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "joblib>=1.5.2",
    "jose>=1.0.0",
    "kiwisolver>=1.4.9",
    "lazy-loader>=0.4",
    "librosa>=0.11.0",
    "llvmlite>=0.45.1",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.3",
    "matplotlib>=3.10.7",
    "mdurl>=0.1.2",
    "msgpack>=1.1.2",
    "multidict>=6.7.0",
    "numba>=0.62.1",
    "numpy>=2.3.5",
    "packaging==25.0",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "pip>=25.3",
    "platformdirs>=4.5.0",
    "pooch>=1.8.2",
    "postgrest>=2.24.0",
    "pretty-errors>=1.2.25",
    "propcache>=0.4.1",
    "psycopg2>=2.9.11",
    "pyasn1>=0.6.1",
    "pycparser>=2.23",
    "pycryptodome>=3.23.0",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings>=2.12.0",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "pyparsing>=3.2.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "realtime>=2.24.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "rich-toolkit>=0.16.0",
    "rignore>=0.7.6",
    "rsa>=4.9.1",
    "ruff==0.14.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "sentry-sdk>=2.46.0",
    "setuptools>=80.9.0",
    "shellingham>=1.5.4",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "soundfile>=0.13.1",
    "soxr>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "sqlmodel==0.0.27",
    "storage3>=2.24.0",
    "strenum>=0.4.15",
    "supabase>=2.24.0",
    "supabase-auth>=2.24.0",
    "supabase-functions>=2.24.0",
    "tensorflow>=2.20.0",
    "tensorflow-hub>=0.16.1",
    "threadpoolctl>=3.6.0",
    "tomli>=2.3.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
    "typer>=0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3>=2.5.0",
    "uvicorn==0.38.0",
    "watchfiles>=1.1.1",
    "websockets>=15.0.1",
    "yarl>=1.22.0",
]
[tool.ruff.lint]
extend-select = ["I"]
